refactor `addNote55` to own file...
breaking compile link to `.vue` files
