+++ packages/vue/src/components/Edit/ViewableDataInputForm/DataInputForm.vue
@@ -104,7 +104,8 @@ import _ from 'lodash';
 import { log } from 'util';
 import Vue from 'vue';
 import { Component, Prop, Watch } from 'vue-property-decorator';
-import { addNote55, addTagToCard, getCourseTagStubs } from '../../../db/courseDB';
+import { addTagToCard, getCourseTagStubs } from '../../../db/courseDB';
+import { addNote55 } from '@/db/courseAPI';
 import SkldrVue from '../../../SkldrVue';
 import AudioInput from './FieldInputs/AudioInput.vue';
 import ImageInput from './FieldInputs/ImageInput.vue';
